% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{construct-str}
\alias{construct-str}
\alias{ic_construct_context_str}
\alias{ic_construct_expression_str}
\title{Construct part of printed string}
\usage{
ic_construct_context_str(context)

ic_construct_expression_str(
  deparsed_expression,
  value,
  peeking.function,
  max.lines
)
}
\arguments{
\item{context}{An object returned by \code{ic_get_context()}.}

\item{deparsed_expression}{The deparsed expression (if present) on which \code{ic()} was called. If
missing (default value), only evaluation context is printed.}

\item{value}{The result of evaluating \code{deparsed_expression}. If expression is missing (default
value), this argument should also be missing.}

\item{peeking.function}{A function to use to transform \code{x} to string.}

\item{max.lines}{A number of lines that the output will be truncated to.}
}
\value{
A part of final string value.
}
\description{
Helper functions when constructing final string for output of \code{ic()}.
}
\section{Functions}{
\itemize{
\item \code{ic_construct_context_str}: Construct context string.

\item \code{ic_construct_expression_str}: Construct expression string.
}}

\keyword{internal}
