% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{enable-disable}
\alias{enable-disable}
\alias{ic_enable}
\alias{ic_disable}
\title{Enable or disable \code{ic()}}
\usage{
ic_enable()

ic_disable()
}
\value{
Returns the old value of the option, invisibly.
}
\description{
These functions enable or disable the \code{ic()} function. While disabled \code{ic()} will do nothing
except evaluate and return its input.
}
\details{
These are just convenience wrappers for \code{options(icecream.enabled = TRUE/FALSE)} used to align
the API with the \href{https://github.com/gruns/icecream#miscellaneous}{Python version}.
}
\section{Functions}{
\itemize{
\item \code{ic_enable()}: Enable \code{ic()}.

\item \code{ic_disable()}: Disable \code{ic()}.

}}
