% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_opt}
\alias{get_opt}
\title{Get value of original option or of newer replacement version}
\usage{
get_opt(original, replacement)
}
\arguments{
\item{original}{A string. Name of the original (pre 0.3.0) option.}

\item{replacement}{A string. Name of the new (since 0.3.0) option.}
}
\description{
This funciton exists to handle deprecated options
}
\keyword{internal}
