% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icecream.R
\name{get_source_info}
\alias{get_source_info}
\title{Retrieve source information from the call stack}
\usage{
get_source_info()
}
\value{
A list containing the source file and line from which the parent function of
\code{get_source_info()} was called.
}
\description{
This is adapted from the source code for \code{\link[=traceback]{traceback()}}.
}
\keyword{internal}
