% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopeek.R
\name{ic_autopeek_header}
\alias{ic_autopeek_header}
\title{Get a header of the object peeked at}
\usage{
ic_autopeek_header(object, ...)
}
\arguments{
\item{object}{The object peeked at.}

\item{...}{Other arguments passed to methods.}
}
\description{
Get a header of the object peeked at
}
\details{
This function is used by \code{ic_autopeek} to get a header of the summary of a object.
It should return object's top-level class name and its dimension.
}
