% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_extract_call_stack}
\alias{ic_extract_call_stack}
\title{Extract call stack from the trace}
\usage{
ic_extract_call_stack(trace)
}
\arguments{
\item{trace}{A stack trace resulting from calling \code{\link[rlang:trace_back]{rlang::trace_back()}}.}
}
\value{
A call stack, list of function calls.
}
\description{
This operation is extracted as a separate function because it contains checking for
rlang function due to compatibility issue.
}
\keyword{internal}
