% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ic_get_context}
\alias{ic_get_context}
\title{Get the context from which \code{ic()} was called}
\usage{
ic_get_context()
}
\value{
A list containing the filaname, line number, environment and function name from which
the call to \code{ic()} was made. Any attributes that do not apply (for example, filename and line
number for functions defined in the global environment) will be NULL.
}
\description{
Captures the context (source file location / environment and calling function) from which \code{ic()}
was called and returns a list of information needed to construct printed output.
}
\keyword{internal}
