% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{ic_peek}
\alias{ic_peek}
\title{Peek at value of expression}
\usage{
ic_peek(
  value,
  peeking_function = getOption("icecream.peeking.function"),
  max_lines = getOption("icecream.max.lines")
)
}
\arguments{
\item{value}{The result of evaluating an expression inside the \code{ic()} function.}

\item{peeking_function}{The function used to peek at the value. Default value is set by the
"icecream.peeking.function" option.}

\item{max_lines}{Maximum number of lines printed. Default value is set by the
"icecream.max.lines" option.}
}
\value{
A string to be printed.
}
\description{
This function is a proxy for calling peeking function.
}
\details{
Default value of \code{icecream.peeking.function} is \code{ic_autopeek}. Suggested possible
alternatives are:
\itemize{
\item \code{utils::str}
\item \code{print}
\item \code{head}
\item \code{summary}
\item \code{tibble::glimpse}
}
}
\seealso{
\code{\link[=ic_autopeek]{ic_autopeek()}} \code{\link[utils:str]{utils::str()}} \code{\link[base:print]{base::print()}} \code{\link[utils:head]{utils::head()}} \code{\link[base:summary]{base::summary()}}
\code{\link[tibble:reexports]{tibble::glimpse()}}
}
\keyword{internal}
