% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{ic_predefined_max_lines}
\alias{ic_predefined_max_lines}
\title{Values of max_lines for predefined peeking functions}
\format{
A named numeric vector of 6 elements. Elements are the max_line values for
predefined peeking functions. Names correspond to names of \link{ic_predefined_peeking_functions}
}
\usage{
ic_predefined_max_lines
}
\description{
Default max_lines values for peeking functions handled by default by the \code{ic_peek()}
}
\keyword{internal}
