% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\docType{data}
\name{ic_predefined_peeking_functions}
\alias{ic_predefined_peeking_functions}
\title{List of predefined peeking functions}
\format{
A named list of 6 elements. Elements are functions that take arbitrary R object as
an input and print to the console or return string as an output. Names of the elements
are names of the function without the package name.
}
\usage{
ic_predefined_peeking_functions
}
\description{
Peeking functions handled by default by the \code{ic_peek()}
}
\keyword{internal}
