% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ic_print}
\alias{ic_print}
\title{Print icecream messages}
\usage{
ic_print(ctx)
}
\arguments{
\item{ctx}{A list containing all the information needed to produce the console message. See
\code{\link[=ic_get_context]{ic_get_context()}} and \code{\link[=ic]{ic()}} itself for details of how this is constructed and what's in it.}
}
\value{
The function returns its final output string invisibly.
}
\description{
The printing logic depends on how the \code{ic()} function has been called and what user options are
set.
}
\keyword{internal}
