% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_print_only_context}
\alias{ic_print_only_context}
\title{Print only context when no value is provided}
\usage{
ic_print_only_context(prefix)
}
\arguments{
\item{prefix}{A prefix to use at the start of the icecream message.}
}
\value{
The same as \code{\link{ic_print}}
}
\description{
Print only context when no value is provided
}
\keyword{internal}
