% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlang-reexports}
\alias{rlang-reexports}
\alias{src_loc}
\alias{relish}
\title{Functions re-exported from rlang}
\usage{
src_loc(srcref, dir = getwd())

relish(x, dir = getwd())
}
\description{
Functions re-exported from rlang
}
\section{Functions}{
\itemize{
\item \code{src_loc}: Return source location from a srcref. If no srcref is found, return
\code{""}.

\item \code{relish}: Tidy file paths
}}

