% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_single}
\alias{simplify_single}
\title{Utility function for simplifying single-element lists}
\usage{
simplify_single(x)
}
\arguments{
\item{x}{A list of length at least one.}
}
\value{
If x has length greater than one, it is returned unchanged. If the length is equal to
one, the first element is returned.
}
\description{
Utility function for simplifying single-element lists
}
