% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{User-friendly debug statements}
\usage{
ic(
  ...,
  prefix = get_opt("icecream.prefix", "icecream_prefix"),
  peeking_function = get_opt("icecream.peeking.function", "icecream_peeking_function"),
  max_lines = get_opt("icecream.max.lines", "icecream_max_lines"),
  always_include_context = get_opt("icecream.always.include.context",
    "icecream_always_include_context")
)
}
\arguments{
\item{...}{Any number of expressions, possibly also 0 expressions.}

\item{prefix}{A prefix to use at the start of the icecream message.}

\item{peeking_function}{A function to use to transform \code{x} to string.}

\item{max_lines}{A number of lines that the output will be truncated to.}

\item{always_include_context}{A logical value. Whether or not to include context when no \code{x} is provided.}
}
\value{
If \code{...} is missing, nothing is returned. If \code{...} is a single expression, returns the result
of evaluating it. If \code{...} contains multiple expressions, it returns list with their values in an
according order.
}
\description{
User-friendly debug statements
}
\details{
Function is primarily called for its side effects. It does not change the value of expression
(or expressions) passed to it, but it prints them with their values to the screen. All named parameters
have corresponding options which are their default values. For details, see \code{\link{icecream}}.
}
\examples{
ic()

f <- function(x) x < 0

ic(f(1))

ic(f(-1))

ic(f(12), sum(1:5), 5)

}
