% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{User-friendly debug statements}
\usage{
ic(x)
}
\arguments{
\item{x}{An expression, or nothing}
}
\value{
If \code{x} is an expression, returns the result of evaluating \code{x}. If \code{x} is missing nothing
is returned.
}
\description{
User-friendly debug statements
}
\examples{
f <- function(x) x < 0

ic(f(1))

ic(f(-1))
}
