% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_evaluate_and_print}
\alias{ic_evaluate_and_print}
\title{Evaluate expressions and print them with value}
\usage{
ic_evaluate_and_print(
  quosures,
  prefix,
  peeking_function,
  max_lines,
  always_include_context
)
}
\arguments{
\item{quosures}{A vector of quosures, contains expressions to be evaluated.}

\item{prefix}{A prefix to use at the start of the icecream message.}

\item{peeking_function}{A function to use to transform \code{x} to string.}

\item{max_lines}{A number of lines that the output will be truncated to.}

\item{always_include_context}{A logical value. Whether or not to include context when no \code{x} is provided.}
}
\value{
Evaluated value of expression \code{quosures}.
}
\description{
Evaluate expressions and print them with value
}
\keyword{internal}
