% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_get_context}
\alias{ic_get_context}
\title{Get context of evaluation of \code{ic()} call}
\usage{
ic_get_context(nest_level = 3)
}
\arguments{
\item{nest_level}{An integer. Number of calls to skip in call stack. Calls need to be skipped
because \code{ic()} and other internal functions add calls to the stack, which are not of our
interest. Default value of 3 corresponds to three skips:
\code{ic()} > \code{ic_evaluate_and_print()}/\code{ic_print_only_context()} > \code{ic_get_context()}}
}
\value{
A list of two objects: \code{loc} containing name of the file where \code{ic()} is called
with line and row number if available, or environment of where the calling function definition
is if not available; \code{parent_ref} containing function in which the \code{ic()} is called if available,
\code{NULL} otherwise.
}
\description{
Get context of evaluation of \code{ic()} call
}
\keyword{internal}
