% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ic_print}
\alias{ic_print}
\title{Print icecream messages}
\usage{
ic_print(loc, parent_ref, expression, value)
}
\arguments{
\item{loc}{String detailing function definition location, which may be a source ref (file, line
number and character index) or an environment.}

\item{parent_ref}{The calling function.}

\item{expression}{The deparsed expression (if present) on which \code{ic()} was called.}

\item{value}{The result of evaluating \code{expression}.}
}
\value{
The function is called for its side effect (printing to the console) but it also returns
its output string, invisibly.
}
\description{
The printing logic depends on how the \code{ic()} function has been called and what user options are
set.
}
\keyword{internal}
