% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ic_print}
\alias{ic_print}
\title{Print icecream messages}
\usage{
ic_print(
  prefix,
  context = rlang::missing_arg(),
  deparsed_exprs = rlang::missing_arg(),
  expr_vals = rlang::missing_arg(),
  peeking.function,
  max.lines
)
}
\arguments{
\item{prefix}{A prefix to use at the start of the icecream message.}

\item{context}{An object returned by \code{ic_get_context()}. A context in which \code{ic()} was called. If
missing (default value), no context is included in the ic message.}

\item{deparsed_exprs}{A list of deparsed expressions (if present) on which \code{ic()} was called. If
missing (default value), only evaluation context is printed.}

\item{expr_vals}{A list with the result of evaluating \code{deparsed_exprs}. If the argument is missing
(default value), this argument should also be missing.}

\item{peeking.function}{A function to use to transform \code{x} to string.}

\item{max.lines}{A number of lines that the output will be truncated to.}
}
\value{
The function is called for its side effect (printing to the console) but it also returns
its output string, invisibly.
}
\description{
The printing logic depends on how the \code{ic()} function has been called and what user options are
set.
}
\keyword{internal}
