% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ic_print}
\alias{ic_print}
\title{Print icecream messages}
\usage{
ic_print(
  loc,
  parent_ref,
  deparsed_expression = missing_arg(),
  value = missing_arg()
)
}
\arguments{
\item{loc}{String detailing function definition location, which may be a source ref (file, line
number and character index) or an environment.}

\item{parent_ref}{The calling function.}

\item{deparsed_expression}{The deparsed expression (if present) on which \code{ic()} was called. If
missing (default value), only evaluation context is printed.}

\item{value}{The result of evaluating \code{deparsed_expression}. If expression is missing (default
value), this argument should also be missing.}
}
\value{
The function is called for its side effect (printing to the console) but it also returns
its output string, invisibly.
}
\description{
The printing logic depends on how the \code{ic()} function has been called and what user options are
set.
}
\keyword{internal}
