% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icecream.R
\docType{package}
\name{icecream}
\alias{icecream}
\title{Icecream: Never debug with \code{print()} again}
\section{Options}{

The following options can be used to control behaviour:
\itemize{
\item \code{icecream.enabled}: Boolean. If \code{FALSE}, calls to \code{ic(foo)} simply evaluate and return \code{foo}.
No output is printed.
\item \code{icecream.prefix}: This is printed at the beginning of every line. Defaults to \code{"ic|"}.
\item \code{icecream.include.context}: Boolean. If \code{TRUE}, when calling \code{ic(foo)} the source file:line
or environment will be printed along with the expression and value. This can be useful for more
complicated debugging but produces a lot of output so is disabled by default.
\item \code{icecream.output.function}: Not implemented yet. See the
\href{https://github.com/gruns/icecream#configuration}{configuration} section of the original
project docs for details of what it will do.
\item \code{icecream.arg.to.string.function}: Not implemented yet. See the
\href{https://github.com/gruns/icecream#configuration}{configuration} section of the original
project docs for details of what it will do.
}
}

