% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rlang-reexports}
\alias{rlang-reexports}
\alias{src_loc}
\alias{relish}
\title{Functions re-exported from rlang}
\usage{
src_loc(srcref, dir = getwd())

relish(x, dir = getwd())
}
\arguments{
\item{srcref}{A \code{srcref} object}

\item{dir}{Directory path}

\item{x}{File path}
}
\value{
Character vector containing either a source location (for \code{src_loc()}) or a path (for
\code{relish()}).
}
\description{
These are used to extract and format source references. As they are not part of the public rlang
API they are reimplemented here.
}
\section{Functions}{
\itemize{
\item \code{src_loc()}: Return source location from a srcref. If no srcref is found, return
\code{""}.

\item \code{relish()}: Tidy file paths by removing instances of \code{dir} from \code{x}. Ensures that
we get a relative path for \code{x}.

}}
\keyword{internal}
